move test features to test library and rename MySetup to Fixture
