+++ searchlib/src/tests/fef/rank_program/rank_program_test.cpp
@@ -8,13 +8,11 @@
 #include <vespa/searchlib/fef/test/queryenvironment.h>
 #include <vespa/searchlib/fef/test/plugin/sum.h>
 #include <vespa/searchlib/fef/rank_program.h>
+#include <vespa/searchlib/fef/test/test_features.h>
 
 using namespace search::fef;
 using namespace search::fef::test;
 using namespace search::features;
-
-using vespalib::eval::DoubleValue;
-using vespalib::eval::ValueType;
 
 uint32_t default_docid = 1;
 
@@ -59,78 +57,13 @@ size_t count_const_features(const RankProgram &program) {
     return count(program, [](const LazyValue &value){ return value.is_const(); });
 }
 
-struct ImpureValueExecutor : FeatureExecutor {
-    double value;
-    ImpureValueExecutor(double value_in) : value(value_in) {}
-    bool isPure() override { return false; }
-    void execute(uint32_t) override { outputs().set_number(0, value); }
-};
-
-struct ImpureValueBlueprint : Blueprint {
-    double value;
-    ImpureValueBlueprint() : Blueprint("ivalue"), value(31212.0) {}
-    void visitDumpFeatures(const IIndexEnvironment &, IDumpFeatureVisitor &) const override {}
-    Blueprint::UP createInstance() const override { return Blueprint::UP(new ImpureValueBlueprint()); }
-    bool setup(const IIndexEnvironment &, const std::vector<vespalib::string> &params) override {
-        ASSERT_EQUAL(1u, params.size());
-        value = strtod(params[0].c_str(), nullptr);
-        describeOutput("out", "the impure value");
-        return true;
-    }
-    FeatureExecutor &createExecutor(const IQueryEnvironment &, vespalib::Stash &stash) const override {
-        return stash.create<ImpureValueExecutor>(value);
-    }
-};
-
-struct DocidExecutor : FeatureExecutor {
-    void execute(uint32_t docid) override { outputs().set_number(0, docid); }
-};
-
-struct DocidBlueprint : Blueprint {
-    DocidBlueprint() : Blueprint("docid") {}
-    void visitDumpFeatures(const IIndexEnvironment &, IDumpFeatureVisitor &) const override {}
-    Blueprint::UP createInstance() const override { return Blueprint::UP(new DocidBlueprint()); }
-    bool setup(const IIndexEnvironment &, const std::vector<vespalib::string> &) override {
-        describeOutput("out", "the local document id");
-        return true;
-    }
-    FeatureExecutor &createExecutor(const IQueryEnvironment &, vespalib::Stash &stash) const override {
-        return stash.create<DocidExecutor>();
-    }
-};
-
-struct BoxingExecutor : FeatureExecutor {
-    DoubleValue value;
-    BoxingExecutor() : value(0.0) {}
-    bool isPure() override { return true; }
-    void execute(uint32_t) override {
-        value = DoubleValue(inputs().get_number(0));
-        outputs().set_object(0, value);
-    }
-};
-
-struct BoxingBlueprint : Blueprint {
-    BoxingBlueprint() : Blueprint("box") {}
-    void visitDumpFeatures(const IIndexEnvironment &, IDumpFeatureVisitor &) const override {}
-    Blueprint::UP createInstance() const override { return Blueprint::UP(new BoxingBlueprint()); }
-    bool setup(const IIndexEnvironment &, const std::vector<vespalib::string> &params) override {
-        ASSERT_EQUAL(1u, params.size());
-        defineInput(params[0]);
-        describeOutput("out", "boxed value", FeatureType::object(ValueType::double_type()));
-        return true;
-    }
-    FeatureExecutor &createExecutor(const IQueryEnvironment &, vespalib::Stash &stash) const override {
-        return stash.create<BoxingExecutor>();
-    }
-};
-
-struct MySetup {
+struct Fixture {
     BlueprintFactory factory;
     IndexEnvironment indexEnv;
     BlueprintResolver::SP resolver;
     Properties overrides;
     RankProgram program;
-    MySetup() : factory(), indexEnv(), resolver(new BlueprintResolver(factory, indexEnv)),
+    Fixture() : factory(), indexEnv(), resolver(new BlueprintResolver(factory, indexEnv)),
                 overrides(), program(resolver)
     {
         factory.addPrototype(Blueprint::SP(new ValueBlueprint()));
@@ -139,15 +72,15 @@ struct MySetup {
         factory.addPrototype(Blueprint::SP(new DocidBlueprint()));
         factory.addPrototype(Blueprint::SP(new BoxingBlueprint()));
     }
-    MySetup &add(const vespalib::string &feature) {
+    Fixture &add(const vespalib::string &feature) {
         resolver->addSeed(feature);
         return *this;
     }
-    MySetup &override(const vespalib::string &feature, double value) {
+    Fixture &override(const vespalib::string &feature, double value) {
         overrides.add(feature, vespalib::make_string("%g", value));
         return *this;
     }
-    MySetup &compile() {
+    Fixture &compile() {
         ASSERT_TRUE(resolver->compile());
         MatchDataLayout mdl;
         QueryEnvironment queryEnv(&indexEnv);
@@ -178,14 +111,14 @@ struct MySetup {
     }
 };
 
-TEST_F("require that simple program works", MySetup()) {
+TEST_F("require that simple program works", Fixture()) {
     EXPECT_EQUAL(15.0, f1.add("mysum(value(10),ivalue(5))").compile().get());
     EXPECT_EQUAL(3u, f1.program.num_executors());
     EXPECT_EQUAL(3u, count_features(f1.program));
     EXPECT_EQUAL(1u, count_const_features(f1.program));
 }
 
-TEST_F("require that const features work", MySetup()) {
+TEST_F("require that const features work", Fixture()) {
     f1.add("mysum(value(10),value(5))").compile();
     EXPECT_EQUAL(15.0, f1.get());
     EXPECT_EQUAL(3u, f1.program.num_executors());
@@ -193,7 +126,7 @@ TEST_F("require that const features work", MySetup()) {
     EXPECT_EQUAL(3u, count_const_features(f1.program));
 }
 
-TEST_F("require that non-const features work", MySetup()) {
+TEST_F("require that non-const features work", Fixture()) {
     f1.add("mysum(ivalue(10),ivalue(5))").compile();
     EXPECT_EQUAL(15.0, f1.get());
     EXPECT_EQUAL(3u, f1.program.num_executors());
@@ -201,7 +134,7 @@ TEST_F("require that non-const features work", MySetup()) {
     EXPECT_EQUAL(0u, count_const_features(f1.program));
 }
 
-TEST_F("require that a single program can calculate multiple output features", MySetup()) {
+TEST_F("require that a single program can calculate multiple output features", Fixture()) {
     f1.add("value(1)").add("ivalue(2)").add("ivalue(3)");
     f1.add("mysum(value(1),value(2),ivalue(3))");
     f1.compile();
@@ -216,7 +149,7 @@ TEST_F("require that a single program can calculate multiple output features", M
     EXPECT_EQUAL(6.0, result["mysum(value(1),value(2),ivalue(3))"]);
 }
 
-TEST_F("require that a single executor can produce multiple features", MySetup()) {
+TEST_F("require that a single executor can produce multiple features", Fixture()) {
     f1.add("mysum(value(1,2,3).0,value(1,2,3).1,value(1,2,3).2)");
     EXPECT_EQUAL(6.0, f1.compile().get());
     EXPECT_EQUAL(2u, f1.program.num_executors());
@@ -224,7 +157,7 @@ TEST_F("require that a single executor can produce multiple features", MySetup()
     EXPECT_EQUAL(4u, count_const_features(f1.program));
 }
 
-TEST_F("require that feature values can be overridden", MySetup()) {
+TEST_F("require that feature values can be overridden", Fixture()) {
     f1.add("value(1)").add("ivalue(2)").add("ivalue(3)");
     f1.add("mysum(value(1),value(2),ivalue(3))");
     f1.override("value(2)", 20.0).override("ivalue(3)", 30.0);
@@ -240,7 +173,7 @@ TEST_F("require that feature values can be overridden", MySetup()) {
     EXPECT_EQUAL(51.0, result["mysum(value(1),value(2),ivalue(3))"]);    
 }
 
-TEST_F("require that the rank program can calculate scores for multiple documents", MySetup()) {
+TEST_F("require that the rank program can calculate scores for multiple documents", Fixture()) {
     f1.add("mysum(value(10),docid)").compile();
     EXPECT_EQUAL(3u, count_features(f1.program));
     EXPECT_EQUAL(1u, count_const_features(f1.program));
@@ -250,7 +183,7 @@ TEST_F("require that the rank program can calculate scores for multiple document
     EXPECT_EQUAL(f1.get(1), 11.0);
 }
 
-TEST_F("require that auto-unboxing of const object values work", MySetup()) {
+TEST_F("require that auto-unboxing of const object values work", Fixture()) {
     f1.add("box(value(10))").compile();
     EXPECT_EQUAL(10.0, f1.get());
     EXPECT_EQUAL(2u, f1.program.num_executors());
@@ -258,7 +191,7 @@ TEST_F("require that auto-unboxing of const object values work", MySetup()) {
     EXPECT_EQUAL(3u, count_const_features(f1.program));
 }
 
-TEST_F("require that auto-unboxing of non-const object values work", MySetup()) {
+TEST_F("require that auto-unboxing of non-const object values work", Fixture()) {
     f1.add("box(ivalue(10))").compile();
     EXPECT_EQUAL(10.0, f1.get());
     EXPECT_EQUAL(2u, f1.program.num_executors());
